remove redundant package
