+++ modules/contracts/apps/subscription-contributors/src/error.rs
