+++ modules/contracts/apps/subscription-contributors/src/msg.rs
