+++ modules/contracts/apps/subscription/examples/schema.rs
@@ -1,7 +1,6 @@
 use abstract_subscription::contract::SubscriptionApp;
-use abstract_subscription::msg::{
-    ConfigResponse, StateResponse, SubscriberStateResponse, SubscriptionFeeResponse,
-};
+use abstract_subscription::msg::{StateResponse, SubscriberStateResponse, SubscriptionFeeResponse};
+use abstract_subscription::state::SubscriptionConfig;
 use cosmwasm_schema::{export_schema, export_schema_with_title, remove_schemas, schema_for};
 use cw_asset::{AssetInfo, AssetInfoUnchecked};
 use std::env::current_dir;
@@ -15,7 +14,7 @@ fn main() {
 
     SubscriptionApp::export_schema(&out_dir);
 
-    export_schema(&schema_for!(ConfigResponse), &out_dir);
+    export_schema(&schema_for!(SubscriptionConfig), &out_dir);
     export_schema(&schema_for!(StateResponse), &out_dir);
     // export_schema(&schema_for!(ContributorStateResponse), &out_dir);
     export_schema(&schema_for!(SubscriberStateResponse), &out_dir);
