+++ modules/contracts/apps/subscription/src/error.rs
