+++ modules/contracts/apps/subscription/tests/integration.rs
@@ -1,20 +1,13 @@
 use abstract_core::objects::{gov_type::GovernanceDetails, AccountId};
-use abstract_interface::{Abstract, AbstractAccount, AppDeployer, VCExecFns};
+use abstract_interface::{Abstract, AbstractAccount, AppDeployer, DeployStrategy, VCExecFns};
 use abstract_subscription::{
     contract::{interface::SubscriptionInterface, CONTRACT_VERSION},
     msg::{SubscriptionExecuteMsgFns, SubscriptionInstantiateMsg, SubscriptionQueryMsgFns},
     state::{EmissionType, SubscriptionConfig},
 };
 
-use abstract_subscription_contributors::{
-    contract::APP_VERSION,
-    msg::{ContributorsExecuteMsgFns, ContributorsInstantiateMsg, ContributorsQueryMsgFns},
-    state::ContributorsConfig,
-    ContributorsInterface,
-};
-use abstract_subscription_interface::contributors::msg as contr_msg;
-use abstract_subscription_interface::subscription::msg as subscr_msg;
-use abstract_subscription_interface::{CONTRIBUTORS_ID, SUBSCRIPTION_ID};
+use abstract_subscription::contract::SUBSCRIPTION_ID;
+use abstract_subscription::msg as subscr_msg;
 use cw20::Cw20Coin;
 use cw20_base::contract::Cw20Base;
 use cw_asset::{AssetInfoBase, AssetInfoUnchecked};
@@ -54,42 +47,6 @@ fn deploy_emission(subscibers: &Subscription) -> anyhow::Result<Cw20Base<Mock>> 
         None,
     )?;
     Ok(emission_cw20)
-}
-
-fn enable_contributors(
-    subscibers: &Subscription,
-    cw20_emission: &Cw20Base<Mock>,
-) -> anyhow::Result<ContributorsInterface<Mock>> {
-    let contributors_app = ContributorsInterface::new(CONTRIBUTORS_ID, subscibers.chain.clone());
-    contributors_app.deploy(APP_VERSION.parse()?)?;
-    subscibers.account.install_app(
-        contributors_app.clone(),
-        &ContributorsInstantiateMsg {
-            protocol_income_share: Default::default(),
-            emission_user_share: Default::default(),
-            max_emissions_multiple: Default::default(),
-            emissions_amp_factor: Default::default(),
-            emissions_offset: Default::default(),
-            token_info: cw_asset::AssetInfoBase::Native("test".to_owned()),
-        },
-        None,
-    )?;
-
-    subscibers.account.manager.execute_on_module(
-        SUBSCRIPTION_ID,
-        subscr_msg::ExecuteMsg::from(
-            subscr_msg::SubscriptionExecuteMsg::UpdateSubscriptionConfig {
-                payment_asset: None,
-                factory_address: None,
-                subscription_cost_per_week: None,
-                contributors_enabled: Some(true),
-                subscription_per_week_emissions: Some(EmissionType::IncomeBased(
-                    AssetInfoUnchecked::Cw20(cw20_emission.addr_str()?),
-                )),
-            },
-        ),
-    )?;
-    Ok(contributors_app)
 }
 
 /// Set up the test environment with the contract installed
@@ -133,7 +90,7 @@ fn setup() -> anyhow::Result<Subscription> {
                 monarch: ADMIN.to_string(),
             })?;
 
-    subscription_app.deploy(APP_VERSION.parse()?)?;
+    subscription_app.deploy(CONTRACT_VERSION.parse()?, DeployStrategy::Try)?;
 
     let cw20_addr = cw20.addr_str()?;
     account.install_app(
@@ -172,51 +129,11 @@ fn successful_install() -> anyhow::Result<()> {
     let config = subscription_app.config()?;
     assert_eq!(
         config,
-        ConfigResponse {
-            subscription: SubscriptionConfig {
-                factory_address: payment_cw20.address()?,
-                payment_asset: cw_asset::AssetInfoBase::Cw20(payment_cw20.address()?),
-                subscription_cost_per_week: Decimal::percent(1),
-                subscription_per_week_emissions: EmissionType::None,
-                contributors_enabled: false
-            }
-        }
-    );
-    Ok(())
-}
-
-#[test]
-fn successful_install_with_contributors() -> anyhow::Result<()> {
-    // Set up the environment and contract
-    let subscription = setup()?;
-
-    let cw20_emissions = deploy_emission(&subscription)?;
-    let contributors = enable_contributors(&subscription, &cw20_emissions)?;
-
-    let subscr_config = subscription.subscription_app.config()?;
-    assert_eq!(
-        subscr_config,
         SubscriptionConfig {
-            factory_address: subscription.payment_cw20.address()?,
-            payment_asset: cw_asset::AssetInfoBase::Cw20(subscription.payment_cw20.address()?),
+            factory_address: payment_cw20.address()?,
+            payment_asset: cw_asset::AssetInfoBase::Cw20(payment_cw20.address()?),
             subscription_cost_per_week: Decimal::percent(1),
-            subscription_per_week_emissions: EmissionType::IncomeBased(AssetInfoBase::Cw20(
-                cw20_emissions.address()?
-            )),
-            contributors_enabled: true
-        }
-    );
-
-    let contr_config = contributors.config()?;
-    assert_eq!(
-        contr_config,
-        ContributorsConfig {
-            protocol_income_share: todo!(),
-            emission_user_share: todo!(),
-            max_emissions_multiple: todo!(),
-            emissions_amp_factor: todo!(),
-            emissions_offset: todo!(),
-            token_info: todo!()
+            subscription_per_week_emissions: EmissionType::None,
         }
     );
     Ok(())
