--- modules/packages/apps/subscription/src/contributors/mod.rs
@@ -1,3 +1,0 @@
-pub mod error;
-pub mod msg;
-pub mod state;
